refactor(services): add published state to data structure
