+++ src/app/data/services/branding.json
@@ -1,4 +1,5 @@
 {
+  "published": true,
   "id": "branding",
   "pageTitle": "Branding & System Design",
   "slug": "branding",
