+++ src/app/data/services/content.json
@@ -1,4 +1,5 @@
 {
+  "published": true,
   "id": "content",
   "pageTitle": "Content creation & craft illustrations",
   "slug": "content",
