+++ src/app/data/services/websites.json
@@ -1,4 +1,5 @@
 {
+  "published": true,
   "id": "websites",
   "pageTitle": "Custom-built websites",
   "slug": "websites",
